refactor: move shared middlewares to the common feature
