+++ src/features/users/users.router.ts
@@ -1,10 +1,10 @@
 import Router from '@koa/router';
 import { object, string } from 'yup';
 import { UsersController } from '@/features/users/users.controller';
-import { vaidateRequestBodyMiddleware } from '@/middlewares/validation/validate-request-body.middleware';
-import { vaidateRequestParamsMiddleware } from '@/middlewares/validation/validate-request-params.middleware';
+import { vaidateRequestBodyMiddleware } from '@/features/common/middlewares/validation/validate-request-body.middleware';
+import { vaidateRequestParamsMiddleware } from '@/features/common/middlewares/validation/validate-request-params.middleware';
 import { findUserByIdMiddleware } from '@/features/users/middlewares/find-user-by-id.middleware';
-import { isCurrentUserMiddleware } from '@/middlewares/authorization/is-current-user.middleware';
+import { isCurrentUserMiddleware } from '@/features/common/middlewares/authorization/is-current-user.middleware';
 import { IUpdateUserDto } from '@/interfaces/dtos/users/update-user.dto';
 import { objectId } from '@/utils/yup/custom-schemas/object-id.schema';
